Eww widget appear on click
The previous implementation of the volume, brightness and bluelight
widgets were taking a small (actually pretty negligible) CPU % because
the values are polled though a `value_monitor.sh` script and states
updated within this script too.

Now the widget visibility are modified directly in hyprland.conf to
ensure no resources is wasted on "listening" - we only show them when we
explicitly want to.

This works for 99% of cases because if we modify volume in some other
way, we likely know the volume - the widget does not tell us things we
do not know